feat: mission and tip
