+++ src/components/layouts/default.vue
@@ -2,7 +2,7 @@
 </script>
 
 <template>
-  <div bg="background" h-full w-full relative p-15px>
+  <div bg="background" h-full w-full relative p="x-15px t-15px b-10px">
     <slot />
   </div>
 </template>
